% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{create_venndiagram}
\alias{create_venndiagram}
\title{Create venndiagram}
\usage{
create_venndiagram(
  x,
  category.names,
  filename,
  main = "",
  main.cex = 0.3,
  cat.cex = 0.15,
  height = 480,
  width = 800,
  ...
)
}
\arguments{
\item{x}{A list of vectors (e.g., integers, chars), with each component corresponding 
  to a separate circle in the Venn diagram}

\item{category.names}{Allow specification of category names using plotmath syntax}

\item{filename}{Filename for image output, or if NULL returns the grid object itself}

\item{main}{Character giving the main title of the diagram}

\item{main.cex}{Number giving the cex (font size) of the main title}

\item{height}{Integer giving the height of the output figure in units}

\item{width}{Integer giving the width of the output figure in units}

\item{...}{A series of graphical parameters tweaking the plot. See below for details}
}
\value{
NA
}
\description{
Create venndiagram
}
