% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_mutations}
\alias{format_mutations}
\title{Formatting VCF files}
\usage{
format_mutations(file)
}
\arguments{
\item{file}{VCF file}
}
\value{
Data frame with formatted VCF file
}
\description{
Format VCF files for downstream analysis
}
